The Great Backend and Frontend Merge(tm)
This is a giant commit, but the individual intermediate steps do not make sense for anyone except me, so .. squash it is. Here's what happened:

1. The backend was re-built using a central project template I have. That template is the result of a lot of experience on what is maintainable and "feels good".
2. The frontend was merged into the backend, using pretty much exclusively server-side rendering. Not only does this reduce the amount of frontend resource bandwidth, it also makes maintenance easier.

For users, there is pretty much no difference, except that the stored filters will no longer be read from localStorage (it's written in the changelog).